requirements.txt
antlr4-python3-runtime==4.9.3
cloudpickle==3.0.0
filelock==3.13.1
fsspec==2024.3.1
gym==0.26.2
gym-notices==0.0.8
hydra-core==1.3.2
imageio==2.34.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
-e git+git@github.com:Dongx1aoYang/IsaacGym.git@b35bbacba6ed0201a9658df642137d91827f6f7f#egg=isaacgym&subdirectory=isaacgym/python
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.19.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.0
pillow==10.2.0
PyYAML==6.0.1
scipy==1.10.1
sympy==1.12
torch==2.2.1
torchvision==0.17.1
triton==2.2.0
typing-extensions==4.10.0
zipp==3.18.1
